remove pubs from RtpsParameter
